+++ protected/components/Controller.php
@@ -315,7 +315,7 @@ class Controller extends BaseController
 				// 'visible' => Yii::app()->user->getState('accessBackend') == true && (Yii::app()->user->isSuperAdmin || Yii::app()->user->isContentManager),
 				'visible' => Yii::app()->user->getState('accessBackend') == true && (
 					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'embed', 'action' => (object)['id' => 'admin']]) ||
-					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'lingual', 'action' => (object)['id' => 'admin']]) ||
+					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'lingual', 'action' => (object)['id' => 'admin'], 'module' => (object)['id' => 'i18n']]) ||
 					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'setting', 'action' => (object)['id' => 'admin']]) ||
 					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'registry', 'action' => (object)['id' => 'admin']]) ||
 					HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'request', 'action' => (object)['id' => 'admin']]) ||
@@ -332,8 +332,8 @@ class Controller extends BaseController
 					),
 					//array('label'=>Yii::t('app', 'Faq'), 'url'=>array('/faq/admin'),),
 					array(
-						'label' => Yii::t('app', 'Lingual'), 'url' => array('/lingual/admin'),
-						'visible' => HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'lingual', 'action' => (object)['id' => 'admin']])
+						'label' => Yii::t('app', 'Lingual'), 'url' => array('/i18n/lingual/admin'),
+						'visible' => HUB::roleCheckerAction(Yii::app()->user->getState('rolesAssigned'), (object)['id' => 'lingual', 'action' => (object)['id' => 'admin'], 'module' => (object)['id' => 'i18n']])
 					),
 					array(
 						'label' => Yii::t('app', 'Setting'), 'url' => array('/setting/admin'),