+++ protected/config/params.php
@@ -37,6 +37,9 @@ return array(
 	'apiPassword' => getenv('API_PASSWORD', 'secret'),
 	'jwtSecret' => getenv('JWT_SECRET', 'thisismysecretkey'),
 
+	'noProxyDomain' => getenv('NO_PROXY_DOMAIN', 'openhubd.mymagic.my'),
+	'noProxyUrl' => getenv('NO_PROXY_URL', '//openhubd.mymagic.my'),
+
 	// do not change this once setup or user password in db will not match during login
 	'saltSecret' => getenv('SALT_SECRET', '0000'),