+++ protected/modules/f7/views/form/_form.php
@@ -136,14 +136,6 @@
 		</div>
 	</div>
 
-	<div class="form-group <?php echo $model->hasErrors('type') ? 'has-error' : '' ?>">
-		<?php echo $form->bsLabelEx2($model, 'Survey'); ?>
-		<div class="col-sm-10">
-			<?php echo $form->bsBooleanList($model, 'type'); ?>
-			<?php echo $form->bsError($model, 'type'); ?>
-		</div>
-	</div>
-
 	<div class="form-group <?php echo $model->hasErrors('is_multiple') ? 'has-error' : '' ?>">
 		<?php echo $form->bsLabelEx2($model, 'is_multiple'); ?>
 		<div class="col-sm-10">