+++ changelog.md
@@ -1,5 +1,8 @@
 ## 0.5.218
 - fixed #32: Saving individual model will trigger error repository() not found 
+- fixed #31: cloudflare error when saving module config
+- refactored lingual to an i18n module under yeebase
+- updated f7 to remove hardcode survey in form
 
 ## 0.5.217
 - fixed #24: ACL do not list actions of modules properly